Enable specifying services as endpoints
Currently a k8s service cannot be specified as an ednpoint to
another service, as the PREROUTING chain (and respective DNAT) is
traversed only once, resulting in a destination address to a k8s
service, whose IP does not really exist.

This commit deals with the issue by resolving the endpoints recursively.